[FEAT] Silence ffmpeg
Currently the call to `ffmpeg` is very noisy as it prints a lot of information such as the build configuration.
This change sets the loglevel to 'fatal' which I think is appropriate here. 'quiet' would also be possible to silence everything but this might hide important information when something goes wrong.

I noticed the noisyness when trying out `swaylock-blur` in a terminal which is suboptimal. Also consider that the log output slows down the process ever so slightly which is at least mildly important when locking the screen.